+++ app/controllers/renalware/hd/sessions_controller.rb
@@ -5,7 +5,6 @@ require "collection_presenter"
 
 module Renalware
   module HD
-    # rubocop:disable Metrics/ClassLength
     class SessionsController < BaseController
       include PresenterHelper
       include Renalware::Concerns::Pageable
@@ -16,24 +15,13 @@ module Renalware
         query = sessions_query
         sessions = query
           .call
-          .eager_load(
-            :hospital_unit,
-            :patient,
-            :signed_on_by,
-            :signed_off_by,
-            prescription_administrations: [
-              {
-                prescription: [:medication_route, :drug]
-              },
-              :administered_by,
-              :reason
-            ])
+          .eager_load(:hospital_unit, :patient, :signed_on_by, :signed_off_by)
           .page(page)
           .per(per_page || 15)
         authorize sessions
         presenter = CollectionPresenter.new(sessions, SessionPresenter, view_context)
         @q = query.search
-        render :index, locals: { sessions: presenter }
+        render :index, locals: { patient: patient, sessions: presenter }
       end
 
       def show
@@ -57,14 +45,6 @@ module Renalware
 
       def edit
         session = Session.for_patient(patient).find(params[:id])
-        session.prescription_administrations.each do |a|
-          if a.administrator_authorised?
-            a.administrator_authorisation_token = a.administered_by&.auth_token
-          end
-          if a.witness_authorised?
-            a.witness_authorisation_token = a.witnessed_by&.auth_token
-          end
-        end
         authorize session
         render :edit, locals: locals(session)
       rescue Pundit::NotAuthorizedError
@@ -110,16 +90,8 @@ module Renalware
       def locals(session)
         {
           session: session,
-          patient: patient,
-          prescriptions_to_be_administered_on_hd: prescriptions_on_hd
+          patient: patient
         }
-      end
-
-      def prescriptions_on_hd
-        @prescriptions_on_hd ||= begin
-          prescriptions = patient.prescriptions.includes([:drug]).to_be_administered_on_hd
-          present(prescriptions, Medications::PrescriptionPresenter)
-        end
       end
 
       def sessions_query
@@ -129,13 +101,9 @@ module Renalware
       def session_params
         @session_params ||= begin
           params.require(:hd_session).require(:type)
-          sesh_params = params.require(:hd_session)
-                              .permit(attributes)
-                              .merge(document: document_attributes, by: current_user)
-          (sesh_params[:prescription_administrations_attributes] || {}).each do |_key, hash|
-            hash[:by] = current_user
-          end
-          sesh_params
+          params.require(:hd_session)
+                        .permit(attributes)
+                        .merge(document: document_attributes, by: current_user)
         end
       end
 
@@ -143,11 +111,6 @@ module Renalware
         [:performed_on, :start_time, :end_time,
          :hospital_unit_id, :notes, :dialysate_id,
          :signed_on_by_id, :signed_off_by_id, :type,
-         prescription_administrations_attributes: [
-           :id, :hd_session_id, :prescription_id, :administered, :notes,
-           :administered_by_id, :administrator_authorisation_token,
-           :witnessed_by_id, :witness_authorisation_token, :reason_id
-         ],
          document: []]
       end
 
@@ -162,6 +125,5 @@ module Renalware
         Delayed::Job.enqueue UpdateRollingPatientStatisticsDjJob.new(patient.id)
       end
     end
-    # rubocop:enable Metrics/ClassLength
   end
 end