+++ app/javascript/renalware/index.js
@@ -16,8 +16,10 @@ import { Application } from "stimulus"
 import TestController from "./controllers/test_controller"
 import AnotherTestController from "./controllers/another_test_controller"
 import ToggleController from "./controllers/toggle_controller"
+import HDPrescriptionController from "./controllers/hd/prescription_administration_controller"
 
 const application = Application.start()
 application.register("test", TestController)
 application.register("another_test", AnotherTestController)
 application.register("toggle", ToggleController)
+application.register("hd-prescription-administration", HDPrescriptionController)