+++ app/models/renalware/hd/patient.rb
@@ -8,6 +8,7 @@ module Renalware
       has_one :hd_profile, class_name: "Profile"
       has_one :hd_preference_set, class_name: "PreferenceSet"
       has_many :hd_sessions, class_name: "Session"
+      has_many :prescription_administrations
       scope :with_profile, lambda {
         includes(hd_profile: :hospital_unit)
         .joins(<<-SQL)