+++ app/policies/renalware/hd/prescription_administration_policy.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Renalware
+  module HD
+    class PrescriptionAdministrationPolicy < BasePolicy
+    end
+  end
+end