+++ db/migrate/20190627141751_add_tokens_to_hd_prescription_administrations.rb
@@ -1,8 +1,6 @@
 class AddTokensToHDPrescriptionAdministrations < ActiveRecord::Migration[5.2]
   def change
     within_renalware_schema do
-      # add_column :hd_prescription_administrations, :administrator_authorisation_token, :string
-      # add_column :hd_prescription_administrations, :witness_authorisation_token, :string
       add_column(
         :hd_prescription_administrations,
         :administrator_authorised,