+++ spec/integration/renalware/patients/patients_spec.rb
@@ -115,7 +115,7 @@ describe "Managing patients", type: :request do
       },
       referral: {
         referring_physician_name: Faker::Name.name,
-        referral_date: Faker::Date.backward(14),
+        referral_date: Faker::Date.backward(days: 14),
         referral_type: "Unknown",
         referral_notes: Faker::Lorem.sentence
       },