+++ spec/models/renalware/hd/patient_spec.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "rails_helper"
+
+module Renalware
+  describe HD::Patient, type: :model do
+    it { is_expected.to have_many(:prescription_administrations) }
+  end
+end